Force float64 conversion in Mangle.get_polyids() for ra/dec.
git-svn-id: svn+ssh://sdss3.org/repo/boss/bosslss/trunk/python@122952 b00c9580-4927-0410-8824-811d448ea4c7

+++ mangle.py
@@ -170,8 +170,9 @@ class Mangle:
 
         ra/dec should be numpy arrays in decimal degrees.
         """
-        theta = pi/180. * (90.0-dec)
-        phi = pi/180. * ra
+        # force an upconversion to double, in case ra/dec are float32
+        theta = pi/180. * (90.0-np.float64(dec))
+        phi = pi/180. * np.float64(ra)
         sintheta = sin(theta)
         x0 = sintheta*cos(phi)
         y0 = sintheta*sin(phi)
